// frontend/src/axios/requests/models/Post14154960View.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Post14154960View = {
    apply_url: string;
    content: string;
    created_at: string;
    created_by: string;
    estimate_salary?: number;
    id: string;
    latitude?: number;
    longitude?: number;
    status: string;
    subHead?: string;
    title: string;
    updated_at?: string;
};