// frontend/src/axios/requests/models/Profile2485805610View.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Profile2485805610View = {
    auth_link: string;
    created_at: string;
    description: string;
    id: number;
    logo: string;
    name: string;
    type: string;
    updated_at?: string;
    website: string;
};