// frontend/src/axios/requests/models/Profile442089671View.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Post14154960View } from './Post14154960View';
export type Profile442089671View = {
    auth_link: string;
    created_at: string;
    description: string;
    id: number;
    logo: string;
    name: string;
    posts?: Array<Post14154960View>;
    type: string;
    updated_at?: string;
    website: string;
};